Replaced VersionRegex with a static method in SdkVersion
