+++ src/dotnet-uninstall/Shared/SdkInfo/SdkVersion.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Text.RegularExpressions;
+using static Microsoft.DotNet.Tools.Uninstall.Shared.Exceptions.Exceptions;
 
 namespace Microsoft.DotNet.Tools.Uninstall.Shared.SdkInfo
 {
@@ -15,6 +17,30 @@ namespace Microsoft.DotNet.Tools.Uninstall.Shared.SdkInfo
             Minor = minor;
             Patch = patch;
             Preview = preview;
+        }
+
+        public static SdkVersion From(string versionString)
+        {
+            var regex = new Regex(@"^(?<major>\d+)\.(?<minor>\d+)\.(?<patch>\d+)((\s\-\s|\-)preview(?<preview>\d+))?$");
+            var match = regex.Match(versionString);
+
+            if (!match.Success)
+            {
+                throw new InvalidVersionStringException(versionString);
+            }
+
+            var versionMajorString = match.Groups["major"].Value;
+            var versionMinorString = match.Groups["minor"].Value;
+            var versionPatchString = match.Groups["patch"].Value;
+
+            var versionPreviewString = match.Groups["preview"].Success ? match.Groups["preview"].Value : null;
+
+            var versionMajor = int.Parse(versionMajorString);
+            var versionMinor = int.Parse(versionMinorString);
+            var versionPatch = int.Parse(versionPatchString);
+            var versionPreview = versionPreviewString != null ? int.Parse(versionPreviewString) as int? : null;
+
+            return new SdkVersion(versionMajor, versionMinor, versionPatch, versionPreview);
         }
 
         public override string ToString()
